Add support for electric bikes
Currently on CityBikes API, electric bikes seem to be supported on
4 networks only, as evidenced in these files :

- pybikes/bicing.py
- pybikes/smovengo.py
- pybikes/keolis.py
- pybikes/smartbike.py

It's not clear if it's standard but every one of them use the same
architecture : the 'extra' JSON item of a station contains the two
following sub-items :
- 'has_ebikes' : boolean
- 'ebikes' : integer
Since on some networks the 'has_bikes' value is computed from the
'ebikes' value being 0 or not 0, only this very data is relevant.

This commit add the 'ebikes' data as Station class's attributes as
well as stations table's columns (upgrade database to version 2).
It's copied from remote object if it exists, otherwise it is null.