Improve screen rotation management
In order to keep track of fragment instances after screen rotation,
the "fragTags" attribute is introduced. This array contain 3 String
objects representing the tags of the tab fragments.
Thus, fragment updates are made using tags instead of references.

Moreover, this commit fixes the 'WindowLeaked' error which is thrown
when a screen rotation occurres during the display of the first run
dialog window. The fix involves the new WelcomeDialogFrament class
whose aim is to handle dialog instead of the StationListActivity.
Indeed DialogFragment behavior is told to be more appropriate by
this guide : https://developer.android.com/guide/topics/ui/dialogs.
Despite this trick, the AlertDialog.Builder is copied as is.