Use ViewModel to manage fragments
Since Android 7 it is not possible to put large Serializable
objects on Bundle to save fragment state. This was necessary
during orientation change, for example.

This commit migrates to the official way to handle this :
https://developer.android.com/guide/fragments/saving-state

The StationsListActivity class dones not save stations lists
into SavedState anymore. The lists are computed by ViewModel
instances which are assigned to each fragment and not passed
in the newInstance method anymore.

Additionaly added the onSaveInstanceState method to save the
emptylist TextView during orientation change.