Enable usage of GPS based Location
Previously the lastKnownLocation was based only on Network Provider.
This provider is only active if the Google Play Services are installed.
The new behaviour is to use wether the GPS or the Network depending if
which one is the more recent :
https://github.com/osmdroid/osmdroid/blob/master/osmdroid-android/
src/main/java/org/osmdroid/util/LocationUtils.java

Issue #53