Adapt behaviour to launch intent to other apps
Since Android 11 it is necessary to declare the need
to query other apps. This declaration is due to
StationActivity which retrieves the number of installed
apps filtering map intents.

As OpenBikeSharing is not available anymore on stores,
the code to trigger markets is useless, thus removed.

Finaly, use recommanded method queryIntentActivities
for Android 13 and above.