Migrate from AsyncTask to ExecutorService
Until now, downloads was done in AsyncTask in StationsListActivity.

This is replaced by a new class JSONDownloadRunnable, which is
executed on a dedicated thread by the Java Executor API.
At the end of the download logic, call onDownloadResultCallback()
of the class which have submitted the task.

This way, the task can be submitted from MapActivty or the widget.